+++ src/extension/background-script/actions/lnurl/withdraw.ts
@@ -1,0 +1,21 @@
+import utils from "../../../../common/lib/utils";
+
+import type { Message, LNURLDetails } from "../../../../types";
+
+async function withdrawWithPrompt(
+  message: Message,
+  lnurlDetails: LNURLDetails
+) {
+  try {
+    const response = await utils.openPrompt({
+      origin: message.origin,
+      type: "lnurlWithdraw",
+      args: { ...message.args, lnurlDetails },
+    });
+    return response; // { data: ... }
+  } catch (e) {
+    return { error: e instanceof Error ? e.message : e };
+  }
+}
+
+export default withdrawWithPrompt;
