+++ src/extension/background-script/actions/webln/index.ts
@@ -1,5 +1,6 @@
 import { sendPaymentOrPrompt } from "./sendPaymentOrPrompt";
 import keysendOrPrompt from "./keysendOrPrompt";
 import signMessageOrPrompt from "./signMessageOrPrompt";
+import lnurl from "./lnurl";
 
-export { sendPaymentOrPrompt, keysendOrPrompt, signMessageOrPrompt };
+export { sendPaymentOrPrompt, keysendOrPrompt, signMessageOrPrompt, lnurl };
