+++ src/extension/background-script/actions/webln/lnurl.ts
@@ -1,0 +1,8 @@
+import lnurl from "../lnurl";
+
+/*
+For now this is simple an alias for ../lnurl
+But this function is called from webln. We can add certain permissions here later
+*/
+
+export default lnurl;
